can delete all tabs
